% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devrcap-package.R
\docType{package}
\name{devrcap-package}
\alias{devrcap}
\alias{devrcap-package}
\title{devrcap: Mastering Software Development in R (Capstone)}
\description{
Collect Code and data for the capstone assignments
    of the Coursera specialization "Mastering Software Development in
    R."
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CorradoLanera/devrcap}
  \item Report bugs at \url{https://github.com/CorradoLanera/devrcap/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@gmail.com}

}
\keyword{internal}
