% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean NOAA data frame}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{A raw data frame of NOAA data}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Clean NOAA data frame
}
\examples{
data(noaa)
tidy_noaa <- eq_location_clean(noaa)
}
